requirements.txt
asgiref==3.7.2
bitarray==2.8.1
cffi==1.15.1
cryptography==41.0.3
cytoolz==0.12.2
Django==4.2.4
djangorestframework==3.14.0
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
hexbytes==0.3.1
jwt==1.3.1
parsimonious==0.9.0
pycparser==2.21
pycryptodome==3.18.0
pytz==2023.3
regex==2023.8.8
rlp==3.0.0
sqlparse==0.4.4
toolz==0.12.0
tzdata==2023.3
